feat: Option to hide tags that are unused in map
A common problem with OSM is the overload of information that beginners get. Most maps don't use
all of the tags, so this adds an optional button to auto-disable any tags that are not used in your map selection.
This prevents the "click-then-squint" pattern of trying to see what, if anything, changed in the map